refactor: Modularize CLI interface and improve code organization
- Split CLI app into separate modules: commands.py, menu.py, and api_key_manager.py
- Improve separation of concerns in CLI implementation
- Enhance code readability and maintainability
- Remove .vscode settings file
- Update import statements and module references
- Add type hints and docstrings to new modules
