+++ llm_key_points/adapters/repositories/json_dataset_repository.py
@@ -3,10 +3,9 @@ JSON file implementation of the dataset repository.
 """
 
 from pathlib import Path
-from typing import List, Dict, Any
 
-from ...core.interfaces.repositories import DatasetRepository
 from ...core.entities.dataset import Dataset
+from ...core.interfaces.repositories import DatasetRepository
 
 
 class JsonDatasetRepository(DatasetRepository):
